refactor : 게임을 시작할 때 html의 element를 숨기는 함수 추가 <description> html에서 style을 적용하는 것이 아닌, 게임이 시작할 때 자동으로 display:none이 되고, 그 후에 필요할 때 display:block으로 만드는 것이 좋을 것 같다고 생각해서 함수를 따로 만들었고, element라는 공통적인 특성이 있어서 elements폴더에 모아놨다. 이렇게 display:none을 html에서만 하는 것이 아닌 게임 시작 시 초기화해주는 이유는 다시하기 버튼을 만들어보려고 하기 때문인데, 다시하기 버튼을 누르면 게임이 초기화되고 다시 시작할 수 있어야 하기 때문이다.
