+++ index.html
@@ -17,13 +17,13 @@
           <input id="car-names-input" type="text" />
           <button id="car-names-submit">확인</button>
         </div>
-        <div id="racing-count-container" style="display: none">
+        <div id="racing-count-container">
           <h4>시도할 횟수를 입력해주세요.</h4>
           <input id="racing-count-input" type="number" />
           <button id="racing-count-submit">확인</button>
         </div>
       </div>
-      <div id="result-container" style="display: none">
+      <div id="result-container">
         <h4>📄 실행 결과</h4>
         <ul id="result-list" style="padding: 0"></ul>
       </div>
