+++ src/init.js
@@ -1,0 +1,6 @@
+import {hideCounterElement, hideResultElement} from './elements/hideElement.js';
+
+export default function initGame() {
+  hideCounterElement();
+  hideResultElement();
+}
