+++ src/play.js
@@ -1,16 +1,22 @@
 import RacingCarGame from './index.js';
 import checkInputValid from './inputvalid.js';
 import {carElements} from './utils.js';
+import initGame from './init.js';
+import {showCounterElement, showResultElement} from './elements/showElement.js';
 
 const racingGame = new RacingCarGame();
 const {
-  timesWrapper,
-  resultWrapper,
   carNameSubmitBtn,
   timesSubmitBtn,
   carNameInput,
   timesInput,
 } = carElements;
+
+const gameStart = () => {
+  initGame();
+};
+
+gameStart();
 
 const onSubmitCarName = (e) => {
   if (!checkInputValid(e.target, carNameInput.value)) {
@@ -18,7 +24,7 @@ const onSubmitCarName = (e) => {
   }
   console.log(racingGame);
   racingGame.getPlayingCarList(carNameInput.value);
-  timesWrapper.style.display = 'block';
+  showCounterElement();
   carNameSubmitBtn.removeEventListener('click', onSubmitCarName);
 
   return timesSubmitBtn.addEventListener('click', onSubmitRacingTimes);
@@ -29,7 +35,7 @@ const onSubmitRacingTimes = (e) => {
     return (timesInput.value = '');
   }
   racingGame.gamePlay(timesInput.value);
-  resultWrapper.style.display = 'block';
+  showResultElement();
   timesSubmitBtn.removeEventListener('click', onSubmitRacingTimes);
 };
