+++ src/progress.js
@@ -1,4 +1,4 @@
-import {createResultList} from './element.js';
+import {createResultList} from './elements/createElement.js';
 
 export default function showOnGameProgress(carsArray) {
   createResultList(carsArray);
