+++ src/utils.js
@@ -1,6 +1,4 @@
 export const carElements = {
-  timesWrapper: document.body.querySelector('#racing-count-container'),
-  resultWrapper: document.body.querySelector('#result-container'),
   carNameSubmitBtn: document.body.querySelector('#car-names-submit'),
   timesSubmitBtn: document.body.querySelector('#racing-count-submit'),
   carNameInput: document.body.querySelector('#car-names-input'),
