+++ src/winner.js
@@ -1,4 +1,4 @@
-import {createWinnerElement} from './element.js';
+import {createWinnerElement} from './elements/createElement.js';
 
 const pickWinner = (carsArray) => {
   const maxLength = findMaxLengthCar(carsArray);
